pyproject.toml
[tool.poetry]
name = "telebot"
version = "0.1.0"
description = ""
authors = ["kremniland <kremnilandk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.4"}
psycopg2 = "^2.9.5"
ipython = "^8.10.0"
asyncio = "^3.4.3"
pillow = "^9.4.0"
requests = "^2.28.2"
asyncpg = "^0.27.0"
alembic = "^1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
